# expression, types: remove the useage of "evalExprToXXX" functions