fix #4405 

before this pr:

```
BenchmarkEvalInt-8              50000000                28.9 ns/op             0 B/op          0 allocs/op
BenchmarkEvalReal-8             50000000                30.5 ns/op             0 B/op          0 allocs/op
BenchmarkEvalString-8           50000000                34.6 ns/op             0 B/op          0 allocs/op
PASS
ok      github.com/pingcap/tidb/expression      4.825s
```

after this pr:
```
BenchmarkEvalInt-8              200000000                6.40 ns/op            0 B/op          0 allocs/op
BenchmarkEvalReal-8             200000000                6.72 ns/op            0 B/op          0 allocs/op
BenchmarkEvalString-8           100000000               15.3 ns/op             0 B/op          0 allocs/op
PASS
ok      github.com/pingcap/tidb/expression      5.554s
```